split new version into second file
